refactor: separate schema utils from validator into their own module (#110)
