+++ src/hca_validation/schema_utils/__init__.py
@@ -1,0 +1,7 @@
+"""
+HCA Validation Tools - Schema Utils Module
+
+This module provides utilities for interacting with the HCA schema.
+"""
+
+from .schema_utils import get_entity_class_name, get_class_identifier_name
