+++ src/hca_validation/validator/__init__.py
@@ -4,4 +4,4 @@ HCA Validation Tools - Validator Module
 This module provides reusable validation components for HCA data validation.
 """
 
-from .validator import get_entity_class_name, validate, validate_id_uniqueness
+from .validator import validate, validate_id_uniqueness
